+++ lib/src/auth/cert_client.dart
@@ -1,3 +1,3 @@
-export 'cert_client_stub.dart'
+export 'cert_client_io.dart'
     if (dart.library.io) 'cert_client_io.dart'
-    if (dart.library.html) 'cert_client_web.dart';
+    if (dart.library.js_interop) 'cert_client_web.dart';