+++ lib/src/auth/cluster.dart
@@ -1,3 +1,3 @@
-export 'cluster_stub.dart'
+export 'cluster_io.dart'
     if (dart.library.io) 'cluster_io.dart'
-    if (dart.library.html) 'cluster_web.dart';
+    if (dart.library.js_interop) 'cluster_web.dart';