+++ lib/src/auth/cluster_web.dart
@@ -7,10 +7,10 @@ import 'package:kuberneteslib/src/helpers/uint8list_converter.dart';
 import '../cluster/cluster.dart';
 import '../cluster/config.dart';
 import '../cluster/user.dart';
-import 'bearer_client_stub.dart'
+import 'bearer_client_io.dart'
     if (dart.library.io) 'bearer_client_io.dart'
     if (dart.library.html) 'bearer_client_web.dart';
-import 'cert_client_stub.dart'
+import 'cert_client_io.dart'
     if (dart.library.io) 'cert_client_io.dart'
     if (dart.library.html) 'cert_client_web.dart';
 import 'cluster_auth_client.dart';