+++ lib/src/status/deployment/deployment_condition.g.dart
@@ -8,12 +8,10 @@ part of 'deployment_condition.dart';
 
 DeploymentCondition _$DeploymentConditionFromJson(Map<String, dynamic> json) =>
     DeploymentCondition(
-      lastTransitionTime: json['lastTransitionTime'] == null
-          ? null
-          : DateTime.parse(json['lastTransitionTime'] as String),
-      lastUpdateTime: json['lastUpdateTime'] == null
-          ? null
-          : DateTime.parse(json['lastUpdateTime'] as String),
+      lastTransitionTime:
+          _lastTransitionTimeFromJson(json['lastTransitionTime'] as String?),
+      lastUpdateTime:
+          _lastUpdateTimeFromJson(json['lastUpdateTime'] as String?),
       message: json['message'] as String?,
       reason: json['reason'] as String?,
       status: json['status'] as String?,
@@ -23,9 +21,10 @@ DeploymentCondition _$DeploymentConditionFromJson(Map<String, dynamic> json) =>
 Map<String, dynamic> _$DeploymentConditionToJson(
         DeploymentCondition instance) =>
     <String, dynamic>{
-      if (instance.lastTransitionTime?.toIso8601String() case final value?)
+      if (_lastTransitionTimeToJson(instance.lastTransitionTime)
+          case final value?)
         'lastTransitionTime': value,
-      if (instance.lastUpdateTime?.toIso8601String() case final value?)
+      if (_lastUpdateTimeToJson(instance.lastUpdateTime) case final value?)
         'lastUpdateTime': value,
       if (instance.message case final value?) 'message': value,
       if (instance.reason case final value?) 'reason': value,